+++ lib/widgets/charts/dashboard_habits_chart.dart
@@ -1,4 +1,5 @@
 import 'dart:core';
+import 'dart:math';
 
 import 'package:beamer/beamer.dart';
 import 'package:collection/collection.dart';
@@ -324,13 +325,19 @@ class _HabitChartLineState extends State<HabitChartLine> {
                     children: [
                       const SizedBox(width: 70),
                       ...results.map((res) {
-                        return ClipRRect(
-                          borderRadius: BorderRadius.circular(2),
-                          child: Container(
-                            height: 25,
-                            width: (MediaQuery.of(context).size.width - 300) /
-                                days,
-                            color: colorFromCssHex(res.hexColor),
+                        return Flexible(
+                          child: ClipRRect(
+                            borderRadius: BorderRadius.circular(2),
+                            child: Container(
+                              margin: EdgeInsets.only(
+                                left: min(
+                                  MediaQuery.of(context).size.width / 6 / days,
+                                  8,
+                                ),
+                              ),
+                              height: 25,
+                              color: colorFromCssHex(res.hexColor),
+                            ),
                           ),
                         );
                       }),