+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed:
 - Bring back index creation in journal database
+- Fix habit success indicator width
 
 ## [0.8.176] - 2022-10-16
 ### Fixed: